requirements.txt
click>=8.1
pandas>=2.0
openpyxl>=3.1
pyyaml>=6.0
rich>=13.0
pytest>=7.0
